Guppy6 and minimap2
This commit changes the minimap2 used for read alignment. Instead of
mappy the Guppy integrated minimap2 is used, this doesn't allow the use
of reference updating or PAF output of alignemnts on this commit.

Reference genome is given by providing the "align_ref" field to the
`caller_settings' table in the config TOML file and the
`conditions.referreference' field is ignored.

In addtion we moved to using the secure gRPC port so require the command
line param `--port 9502'.